+++ crates/amaru-pure-stage/src/lib.rs
@@ -53,5 +53,7 @@ pub use trace_match::{
     TraceMatch, assert_trace_contains, assert_trace_does_not_contain, assert_trace_match, tm_add_stage,
     tm_external_effect, tm_external_effect_match, tm_input, tm_send, tm_state, tm_terminate, tm_terminated,
 };
-pub use types::{BLACKHOLE_NAME, BoxFuture, Name, OrTerminateWith, SendData, TryInStage, Void, err, warn};
+pub use types::{
+    BLACKHOLE_NAME, BoxFuture, Name, OrTerminateWith, PRIORITY_MAILBOX_SIZE, SendData, TryInStage, Void, err, warn,
+};
 pub use typetag;